Richard,
First _draft_ of what we said about switching on Monday (also in SVN
along with a more complete POMDP bibliography -- bib. still needs some
editing). If you need more details in a particular direction let me
know. I have a writeup of the plangraph I can pull out of an old
article I wrote if you are desperate for material. I can try and flesh
this out in the last minute, but if you give me until the end of the
month I can obviously be a lot more thorough and detailed.

Okay, I will read this again tomorrow and get back to you with
improvements. I have not drafted the FSC (finite state controller)
POMDP paragraphs yet. That tasks requires much more careful phrasing.

If you "SVN up" your report I can try and integrate this text more
closely with your vocab. etc.

Incidentally, I was thinking of phrasing "criticality constraint" in
terms of PDDL3 constraints, however that will require some effort on
my part to express in this document, and agreement from Michael-B to
implement support for them in his system.

cheers,
Charles


Generally it is not possible to evaluate the expected performance of a
continual planner without performing reasoning that is analogous to
decision-theoretic planning -- i.e., without invoking the planning
phase of the replanner at projected states. Similarly, the
computational expense of evaluating the likelihood of a continual
planner causing the robot to traverse into a dead end -- i.e., a state
from which the goal cannot be achieved -- is of the same order as
decision-theoretic planning. Despite these gloomy observations, in
practice we do have pragmatic strategies for switching between
continual and decision-theoretic planning. Indeed, we shall employ two
such {\em switching} strategies in CogX.


Each strategy requires that the preliminary planning phase of
continual planning is complete, and that the resultant {\em serial
plan} is available. This plan corresponds to a sequence of actions
that when executed can achieve the goal with non-zero probability. Our
first switching strategy, called {\em negligible-uncertainty}
switching, is the simplest. Here, if the proposed serial plan achieves
the goal with near certainty, then no more-sophisticated planning is
undertaken. This covers situations where there is very little
uncertainty in the task at hand. For example, navigation actions
presented to the CogX planning subarchitecture are more-or-less
deterministic, and hence so is the outcome of short-term navigational
plans.

Our second switching strategy, called {\em denied-state} switching,
requires the notion of a {\em criticality constraint}. Each such
constraint is represented by a proposition that is labelled as
critically requiring that its assignment to a certain truth value be
maintained. For example, it might be acceptable for the robot to carry
weet-bix into a bedroom, however weet-bix must never be in a
bathroom. Therefore, if the only way to the bedroom is to traverse a
bathroom, because the robot can drop what it is carrying, any plan to
deliver the weet-bix to the bedroom can cause a violation of the
critical ``no weet-bix in the bathroom'' constraint. We address this
problem by prohibiting the continual planner from sequencing actions
that would violate a criticality constraint. Moreover, for any state
which can arise if we execute the serial plan, we perform {\em
plangraph} {\em reachabilioty analysis}~\cite{blum:furst:97} to ensure
there is sufficient reason to believe the goal remains reachable from
that state while respecting criticality. If reachability is denied for
any such projected state, then the planning subarchitecture switches
to using a decision-theoretic planner which is able to carefully
trade-off the costs of violating criticality constraints against
rewards associated with goal achievement.