Richard,
It became obvious this afternoon that I was not going to come up with
a coherent AVS domain and problem definition before the deadline. So I
modified the IPC-5 tireworld to given an instantiation of my
grammar. Similarly, I don't have time between now and the deadline to
write up a formal semantics. Would take me at least a week to get a
draft I was happy with.

cheers,
Charles