Hi Folks,
You will find in "sketch.example.pddl" a sketch of the switching
planner interacting with this problem and domain. Please get back to me
if there are any problems, technical or otherwise, with the material
so far -- i.e., I could repair it between now and Vienna.

Looking forward to getting some constructive comments :))

cheers,
Charles