OKay, I still can't get DORA to compile.
Rather than fight with that, I merged the ../cassandra/ DP with this source. So now there is rudimentary DP functionality all in one place. This should do for the code camp. Now it is just a matter of getting this to play with DORA. I really wanted to test the trajectory expansion code before I tested DP with DORA. But that will not happen now.

Obvious large TODO list, but I think the priority now should be getting this working with Moritz's stuff before I head to Greece on Sunday.

 - Charles