Change lots of behaviour regarding placeholders.
The idea is to remove placeholders when they are no longer interesting. That is, when we would no longer place a new placeholder at a certain position, we should also remove any existing placeholders at that position. In practise, this usually means that we will remove palceholders that are no longer on frontiers, or frontiers which we can not move to.

As of now, every time we try to add new placeholders, we also remove the ones that are no longer interesting. This also means that we can remove some of the old checks that tried to keep the amount of placeholders in check.

--max-frontier-distance is gone
--min-frontier-length is gone
--hyp-path-length is gone
--min-placeholder-to-wall-distance is gone