Hack to include kinect data in the combined local gridmaps used for viewcone generation.
I'm not entirely happy with this as it would be nicer to not have to couple SpatialControl and LocalMapManager so tightly. It would also be nice not having to send map data over an interface. The maps are too tightly tied to SpatialControl as it is however, and duplicating the effort of mapping the kinect data to the local maps is too expensive.