Make sure the gridmaps are not updated when the pantilt does not have a tilt of -45 degrees.
A big warning:
After the pantilt command finishes and we update the gridmaps, we still some
times get lots of garbage. This seems to be a race where we get slightly old
data. For now, we sleep for one second before we update the gridmaps again so
that we get fresh data.

A slightly bigger warning:
When we don't update the gridmaps, we don't update the cure obstacle map. This
means we will run into stuff if we're moving.